+++ trunk/code/otf_analyzer/include/MPIAnalysis.hpp
@@ -34,21 +34,21 @@ namespace cdm
 
         typedef struct
         {
-            uint32_t processID;
+            uint64_t processID;
             uint32_t nodeID;
         } ProcessNodePair;
 
     private:
         typedef std::map<uint32_t, uint32_t> TokenTokenMap;
-        typedef std::map<uint32_t, GraphNode*> IdNodeMap;
-        typedef std::map<uint32_t, IdNodeMap > RemoteNodeMap;
+        typedef std::map<uint64_t, GraphNode*> IdNodeMap;
+        typedef std::map<uint64_t, IdNodeMap > RemoteNodeMap;
         typedef std::map<GraphNode*, ProcessNodePair> ReverseRemoteNodeMap;
     public:
 
         typedef struct
         {
             MPI_Comm comm;
-            std::set<uint32_t> procs;
+            std::set<uint64_t> procs;
         } MPICommGroup;
         
         enum MPIEdgeDirection
@@ -61,15 +61,15 @@ namespace cdm
         {
             MPIEdgeDirection direction;
             GraphNode *localNode;
-            uint32_t remoteNodeID; // remote node ID
-            uint32_t remoteProcessID; // remote process ID
+            uint64_t remoteNodeID; // remote node ID
+            uint64_t remoteProcessID; // remote process ID
         } MPIEdge;
 
         typedef struct
         {
-            //uint32_t prevProcessID;
-            uint32_t processID;
-            //uint32_t nextProcessID;
+            //uint64_t prevProcessID;
+            uint64_t processID;
+            //uint64_t nextProcessID;
             uint32_t nodeStartID;
             uint32_t nodeEndID;
         } CriticalPathSection;
@@ -77,25 +77,25 @@ namespace cdm
         typedef std::map<GraphNode*, CriticalPathSection> CriticalSectionsMap;
         typedef std::vector<CriticalPathSection> CriticalSectionsList;
 
-        typedef std::map<uint32_t, MPIEdge> MPIIdEdgeMap;
-        typedef std::map<uint32_t, MPIIdEdgeMap> MPIRemoteEdgeMap;
-        typedef std::map<uint32_t, MPICommGroup > MPICommGroupMap;
+        typedef std::map<uint64_t, MPIEdge> MPIIdEdgeMap;
+        typedef std::map<uint64_t, MPIIdEdgeMap> MPIRemoteEdgeMap;
+        typedef std::map<uint64_t, MPICommGroup > MPICommGroupMap;
 
         MPIAnalysis(uint32_t mpiRank, uint32_t mpiSize);
         virtual ~MPIAnalysis();
 
         uint32_t getMPIRank() const;
         uint32_t getMPISize() const;
-        uint32_t getMPIRank(uint32_t processId) const;
-        uint32_t getMPIRank(uint32_t processId, const MPICommGroup &commGroup) const;
-        void setMPIRank(uint32_t processId, uint32_t rank);
-        void setMPICommGroupMap(uint32_t group, uint32_t numProcs, const uint32_t *procs);
+        uint32_t getMPIRank(uint64_t processId) const;
+        uint32_t getMPIRank(uint64_t processId, const MPICommGroup &commGroup) const;
+        void setMPIRank(uint64_t processId, uint32_t rank);
+        void setMPICommGroupMap(uint32_t group, uint32_t numProcs, const uint64_t *procs);
         void createMPICommunicatorsFromMap();
         const MPICommGroup& getMPICommGroup(uint32_t group) const;
 
         void addRemoteMPIEdge(GraphNode *localNode, uint32_t remoteNodeID,
-            uint32_t remoteProcessID, MPIEdgeDirection direction);
-        bool getRemoteMPIEdge(uint32_t remoteNodeId, uint32_t remoteProcessId,
+            uint64_t remoteProcessID, MPIEdgeDirection direction);
+        bool getRemoteMPIEdge(uint32_t remoteNodeId, uint64_t remoteProcessId,
             MPIEdge &edge);
 
         ProcessNodePair getRemoteNodeInfo(GraphNode *localNode, bool *valid);
